# backend/services/gamesite-backend/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/leaveRoom/{roomId}":{"post":{"tags":["rest-web-controller"],"summary":"Leaves a room","operationId":"postLeaveRoom","parameters":[{"name":"roomId","in":"path","description":"The id of the room to lave","required":true,"schema":{"type":"string"}}],"responses":{"401":{"description":"Authentication is required"},"400":{"description":"Tried to leave a room that the given player is not part of"},"200":{"description":"Successfully joined room"}}}},"/api/joinRoom/{roomId}":{"post":{"tags":["rest-web-controller"],"summary":"Joins a room. Returns whether or not action was successful","operationId":"postJoinRoom","parameters":[{"name":"roomId","in":"path","description":"The id of the room","required":true,"schema":{"type":"string"}},{"name":"asSpectator","in":"query","description":"Whether player should join as a spectator","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"Successfully joined room","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomJoinResponse"}}}},"401":{"description":"Authentication is required"},"403":{"description":"Not allowed to join the room"}}}},"/api/createRoom":{"post":{"tags":["rest-web-controller"],"summary":"Creates a room","operationId":"postCreateRoom","parameters":[{"name":"maxLobbySize","in":"query","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"gameId","in":"query","required":true,"schema":{"type":"string"}},{"name":"roomName","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully creates a room","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomCreateResponse"}}}},"401":{"description":"Authentication is required"},"400":{"description":"Invalid request. Read body","content":{"text/plain":{}}}}}},"/api/authenticate":{"post":{"tags":["rest-web-controller"],"summary":"Authenticates the user","operationId":"postAuth","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAuthRequest"}}},"required":true},"responses":{"401":{"description":"Bad login or password"},"400":{"description":"Authentication information was not provided"},"200":{"description":"Successful","headers":{"Set-Cookie":{"style":"simple"}}}}}},"/api/roomInfo/{roomId}":{"get":{"tags":["rest-web-controller"],"summary":"Gets information about a room","operationId":"getRoomInfo","parameters":[{"name":"roomId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successful on getting room information","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomInfoResponse"}}}},"401":{"description":"Authentication is required"},"400":{"description":"Invalid request. Read body","content":{"text/plain":{}}},"404":{"description":"Room Id does not exist"}}}},"/api/refreshToken":{"get":{"tags":["rest-web-controller"],"summary":"Refreshes the authentication cookie","operationId":"getRefreshToken","responses":{"200":{"description":"Successful","headers":{"Set-Cookie":{"style":"simple"}}},"401":{"description":"Authentication is required"}}}},"/api/getRoomLists":{"get":{"tags":["rest-web-controller"],"summary":"Get a list of available rooms","operationId":"getRoomLists","responses":{"200":{"description":"Successfully retrieved rooms","content":{"application.json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomListing"}}}}},"401":{"description":"Authentication is required"}}}},"/api/getGames":{"get":{"tags":["rest-web-controller"],"summary":"Get a list of available games","operationId":"getGameList","responses":{"401":{"description":"Authentication is required"},"200":{"description":"Successfully retrieved games","content":{"application.json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomListing"}}}}}}}},"/api/invalidateAuthentication":{"delete":{"tags":["rest-web-controller"],"summary":"Invalidates token and session","operationId":"deleteAuthenticationToken","responses":{"200":{"description":"Successful","headers":{"Set-Cookie":{"style":"simple"}}},"401":{"description":"Authentication is required"}}}}},"components":{"schemas":{"RoomJoinResponse":{"type":"object","properties":{"success":{"type":"boolean"},"isHost":{"type":"boolean"}}},"RoomCreateResponse":{"type":"object","properties":{"success":{"type":"boolean"},"roomId":{"type":"string"}}},"UserAuthRequest":{"type":"object","properties":{"login":{"type":"string"},"password":{"type":"string"}}},"RoomInfoResponse":{"type":"object","properties":{"room":{"$ref":"#/components/schemas/RoomListing"},"isHost":{"type":"boolean"},"joinedRoom":{"type":"boolean"},"isSpectating":{"type":"boolean"}}},"RoomListing":{"type":"object","properties":{"roomId":{"type":"string"},"lobbySize":{"type":"integer","format":"int32"},"maxLobbySize":{"type":"integer","format":"int32"},"spectatorsAmount":{"type":"integer","format":"int32"},"gameType":{"type":"string"},"hostName":{"type":"string"},"inProgress":{"type":"boolean"},"gameStartTime":{"type":"integer","format":"int64"},"roomName":{"type":"string"}}}}}}